move mailbox logic to it's own module
