+++ spec/features/ncr/work_orders/requester_edit_spec.rb
@@ -78,7 +78,7 @@ feature 'Requester edits their NCR work order' do
     click_on "Update"
     proposal = Proposal.last
     expect(proposal.approvers.length).to eq(2)
-    expect(proposal.approvers.second.email_address).to eq(Ncr::ApprovalManager.ba80_budget_mailbox)
+    expect(proposal.approvers.second.email_address).to eq(Ncr::Mailboxes.ba80_budget)
   end
 
   context "proposal changes from BA80 to BA61" do
